LayoutTests/accessibility/ax-value-with-search-expected.txt

This tests that a search field is returns its accessibilityValue correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS searchInput.stringValue is 'AXValue: hello'
PASS searchInput.stringValue === 'AXValue: test'
PASS successfullyParsed is true

TEST COMPLETE
